Added the /proc/stat support for the xavier_config.json configuration file

+++ src/exports/export_struct.h
@@ -16,7 +16,7 @@ struct Process
 struct MeasuredItem
 {
   int id;
-  int measuredValue;
+  double measuredValue;
 };
 struct ExportData
 {
